==> Script/1_Single_cell_pipeline/3_Single_cell_model_training.ipynb <==
from mofapy2.run.entry_point import entry_point
import pandas as pd
import numpy as np
import mofax as mfx

# initialise the entry point
ent = entry_point()

# --- Read data ---
proteomics = pd.read_csv("C:/Users/49152/Downloads/Multi-omics/MOFA/input/1_Single_cell/proteomics_for_MOFA.csv")
transcriptomics = pd.read_csv("C:/Users/49152/Downloads/Multi-omics/MOFA/input/1_Single_cell/transcriptomics_for_MOFA.csv")

# Move "Gene" column into index (like rownames in R)
proteomics_mat = proteomics.set_index("Gene")
transcriptomics_mat = transcriptomics.set_index("Gene")

# --- Reshape to long format ---
proteomics_long = (
    proteomics_mat
    .reset_index()
    .melt(id_vars="Gene", var_name="sample", value_name="value")
    .rename(columns={"Gene": "feature"})
)
proteomics_long["view"] = "Proteomics"
# Make features unique by adding suffix
proteomics_long["feature"] = proteomics_long["feature"].astype(str) + "_prot"

transcriptomics_long = (
    transcriptomics_mat
    .reset_index()
    .melt(id_vars="Gene", var_name="sample", value_name="value")
    .rename(columns={"Gene": "feature"})
)
transcriptomics_long["view"] = "Transcriptomics"
# Make features unique by adding suffix
transcriptomics_long["feature"] = transcriptomics_long["feature"].astype(str) + "_rna"

# --- Combine into one long DataFrame ---
mofa_long = pd.concat([proteomics_long, transcriptomics_long], ignore_index=True)

mofa_long

# Data.frame format
ent.set_data_df(mofa_long, likelihoods = ["gaussian","gaussian"])

ent.set_model_options(factors = 7)

ent.set_train_options(
    convergence_mode = "slow", 
    dropR2 = 0, 
    gpu_mode = True, 
    seed = 1
)

ent.build()

ent.run()

# Save the output
ent.save(outfile="C:/Users/49152/Downloads/Multi-omics/MOFA/output/single_cell_trained_model.hdf5")
